Started implementing tensorflow custom sleeve code.
